samin type stability work

+++ src/Optimization/samin.jl
@@ -134,7 +134,7 @@ function samin(x::Int64)
     return xtest, ftest
 end
 
-function samin(obj_fn, x::Vector{Float64}, lb::Vector{Float64}, ub::Vector{Float64}; nt=5, ns=5, rt=0.5, maxevals=1e6, neps=5, functol=1e-8, paramtol=1e-5, verbosity=1, coverage_ok=0)
+function samin(obj_fn, x::Vector{Float64}, lb::Vector{Float64}, ub::Vector{Float64}; nt=5, ns=5, rt=0.5, maxevals=Int64(1e6), neps=5, functol=1e-8, paramtol=1e-5, verbosity=1, coverage_ok=0)
     n = size(x,1) # dimension of parameter
     #  Set initial values
     nacc = 0 # total accepted trials
